// frontend/src/app/features/tasks/models/task.model.ts
export interface Task {
  id: number;
  title: string;
  description: string | null;
  is_done: boolean;
  created_at: string;
  updated_at: string;
}

export interface PaginatedResponse<T> {
  data: T[];
  links: {
    first: string | null;
    last: string | null;
    prev: string | null;
    next: string | null;
  };
  meta: {
    current_page: number;
    last_page: number;
    per_page: number;
    total: number;
  };
}